prevent unauthorised acces using policy

+++ app/Http/Controllers/StoriesController.php
@@ -31,6 +31,8 @@ class StoriesController extends Controller
      */
     public function create()
     {
+        $this->authorize('update', $story);
+
         $story = new Story;
         return view('stories.create', [
             'story' => $story,
@@ -58,6 +60,7 @@ class StoriesController extends Controller
      */
     public function show(Story $story)
     {
+        $this->authorize('view', $story);
         return view('stories.show', [
             'story' => $story,
         ]);
@@ -72,7 +75,8 @@ class StoriesController extends Controller
     public function edit(Story $story)
     {
         // ? Prevent other users from editing another user's story
-        Gate::authorize('edit-story', $story);
+        // Gate::authorize('edit-story', $story);
+        $this->authorize('update', $story);
         // ?
         return view('stories.edit', [
             'story' => $story,
@@ -101,6 +105,7 @@ class StoriesController extends Controller
      */
     public function destroy(Story $story)
     {
+        $this->authorize('delete', $story);
         $story->delete();
         return redirect()->route('stories.index')
             ->with('status', 'Story Deleted Successfully');
